cmd/erigon, db/kv/backup: add 'erigon db compact'
Moves the datadir walk out of 'integration mdbx_to_mdbx' into
backup.CompactDatadir, so the main binary can compact a datadir too.
