+++ cmd/integration/commands/refetence_db.go
@@ -24,7 +24,6 @@ import (
 	"errors"
 	"fmt"
 	"os"
-	"path/filepath"
 	"strings"
 	"time"
 
@@ -32,7 +31,6 @@ import (
 	"golang.org/x/sync/semaphore"
 
 	"github.com/erigontech/erigon/common"
-	"github.com/erigontech/erigon/common/dir"
 	"github.com/erigontech/erigon/common/log/v3"
 	"github.com/erigontech/erigon/db/datadir"
 	"github.com/erigontech/erigon/db/kv"
@@ -109,7 +107,7 @@ var cmdMdbxToMdbx = &cobra.Command{
 
 		var err error
 		if toChaindata == "" {
-			err = compactDatadir(ctx, logger)
+			err = backup.CompactDatadir(ctx, datadir.Open(datadirCli), logger)
 		} else {
 			from, to := backup.OpenPair(chaindata, toChaindata, dbcfg.ChainDB, 0, logger)
 			err = backup.Kv2kv(ctx, from, to, nil, logger)
@@ -119,89 +117,6 @@ var cmdMdbxToMdbx = &cobra.Command{
 		}
 		return err
 	},
-}
-
-type datadirDB struct {
-	path  string
-	label kv.Label
-}
-
-// caplin/blobs/chaindata is the deepest db a datadir holds.
-const maxDBDepth = 2
-
-// datadirDBs finds the mdbx databases of a datadir. It only descends into the
-// top-level dirs erigon puts a db under - never into snapshots/ or temp/ - and
-// takes the label from that dir, so each copy is sized like the node sizes it.
-func datadirDBs(dirs datadir.Dirs) ([]datadirDB, error) {
-	roots := []datadirDB{
-		{dirs.Chaindata, dbcfg.ChainDB},
-		{filepath.Join(dirs.DataDir, "aura"), dbcfg.ConsensusDB},
-		{dirs.TxPool, dbcfg.TxPoolDB},
-		{dirs.Downloader, dbcfg.DownloaderDB},
-		{dirs.Migrations, dbcfg.MigrationsDB},
-		{dirs.Nodes, dbcfg.SentryDB},
-		{filepath.Join(dirs.DataDir, "caplin"), dbcfg.CaplinDB},
-	}
-	var found []datadirDB
-	for _, root := range roots {
-		if err := findDBs(root.path, root.label, maxDBDepth, &found); err != nil {
-			return nil, err
-		}
-	}
-	return found, nil
-}
-
-func findDBs(path string, label kv.Label, depth int, found *[]datadirDB) error {
-	exists, err := dir.FileExist(filepath.Join(path, "mdbx.dat"))
-	if err != nil {
-		return err
-	}
-	if exists {
-		*found = append(*found, datadirDB{path, label})
-		return nil
-	}
-	if depth == 0 {
-		return nil
-	}
-	entries, err := os.ReadDir(path)
-	if err != nil {
-		if os.IsNotExist(err) {
-			return nil
-		}
-		return err
-	}
-	for _, e := range entries {
-		if !e.IsDir() {
-			continue
-		}
-		if err := findDBs(filepath.Join(path, e.Name()), label, depth-1, found); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func compactDatadir(ctx context.Context, logger log.Logger) error {
-	dirs, l, err := datadir.Open(datadirCli).MustFlock()
-	if err != nil {
-		return err
-	}
-	defer func() {
-		if err := l.Unlock(); err != nil {
-			logger.Error("failed to unlock datadir", "err", err)
-		}
-	}()
-
-	dbs, err := datadirDBs(dirs)
-	if err != nil {
-		return err
-	}
-	for _, db := range dbs {
-		if err := backup.CompactInPlace(ctx, db.path, db.label, logger); err != nil {
-			return fmt.Errorf("compacting %s: %w", db.path, err)
-		}
-	}
-	return nil
 }
 
 var cmdFToMdbx = &cobra.Command{
